requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0